requirements.txt
streamlit>=1.35.0
psycopg2-binary>=2.9.9
pandas>=2.1.0
openpyxl>=3.1.2
xlrd>=2.0.1
